heap_file

+++ src/storage/heap_file.c
@@ -25224,7 +25224,7 @@ heap_log_postpone_heap_append_pages (THREAD_ENTRY * thread_p, const HFID * hfid,
 
   // This append needs to be run on postpone after the commit.
   // First create the log data required.
-  int array_size = heap_pages_array.size ();
+  size_t array_size = heap_pages_array.size ();
   int log_data_size = (DB_ALIGN (OR_HFID_SIZE, PTR_ALIGNMENT) + OR_OID_SIZE + sizeof (int)
                        + array_size * DISK_VPID_ALIGNED_SIZE);
   char *log_data = (char *) db_private_alloc (NULL, log_data_size + MAX_ALIGNMENT);
@@ -25244,7 +25244,7 @@ heap_log_postpone_heap_append_pages (THREAD_ENTRY * thread_p, const HFID * hfid,
   ptr = PTR_ALIGN (ptr, PTR_ALIGNMENT);
 
   // array_size
-  OR_PUT_INT (ptr, array_size);
+  OR_PUT_INT (ptr, (int) array_size);
   ptr += OR_INT_SIZE;
 
   // The array of VPID.
